ch18

+++ ch13 持久层Mybatis框架/crud/src/main/java/com/javaee/mybatis/crud/domain/Teacher.java
@@ -10,7 +10,7 @@ public class Teacher {
     Long id;
     String name;
     String office;
-    Short age;
+    Integer age;
     String email;
     String cellphone;
 }
